remove unused fixtures
