+++ tests/cellfinder_core/test_integration/test_cellfinder.py
@@ -1,27 +1,11 @@
 import subprocess
 import sys
 from pathlib import Path
-from typing import Optional
-
-import pytest
 
 from brainglobe_workflows.cellfinder_core.cellfinder import main
 
 
-@pytest.mark.parametrize(
-    "input_config",
-    [
-        None,
-        "input_config_fetch_GIN",
-        "input_config_fetch_local",
-    ],
-)
-def test_main(
-    input_config: Optional[str],
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-    request: pytest.FixtureRequest,
-):
+def test_main():
     """Test main function for setting up and running cellfinder workflow
 
     Parameters
@@ -35,35 +19,15 @@ def test_main(
     request : pytest.FixtureRequest
         Pytest fixture to enable requesting fixtures by name
     """
-    # monkeypatch to change current directory to
-    # pytest temporary directory
-    # (cellfinder cache directory is created in cwd)
-    monkeypatch.chdir(tmp_path)
 
     # run main
-    if not input_config:
-        cfg = main()
-    else:
-        cfg = main(str(request.getfixturevalue(input_config)))
+    cfg = main()
 
     # check output files exist
     assert Path(cfg._detected_cells_path).is_file()
 
 
-@pytest.mark.parametrize(
-    "input_config",
-    [
-        None,
-        "input_config_fetch_GIN",
-        "input_config_fetch_local",
-    ],
-)
-def test_script(
-    input_config: Optional[str],
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-    request: pytest.FixtureRequest,
-):
+def test_script():
     """Test running the cellfinder worklfow from the command line
 
     Parameters
@@ -77,10 +41,6 @@ def test_script(
     request : pytest.FixtureRequest
         Pytest fixture to enable requesting fixtures by name
     """
-    # monkeypatch to change current directory to
-    # pytest temporary directory
-    # (cellfinder cache directory is created in cwd)
-    monkeypatch.chdir(tmp_path)
 
     # define CLI input
     script_path = (
@@ -93,10 +53,6 @@ def test_script(
         sys.executable,
         str(script_path),
     ]
-    # append config if required
-    if input_config:
-        subprocess_input.append("--config")
-        subprocess_input.append(str(request.getfixturevalue(input_config)))
 
     # run workflow script from the CLI
     subprocess_output = subprocess.run(
@@ -107,20 +63,7 @@ def test_script(
     assert subprocess_output.returncode == 0
 
 
-@pytest.mark.parametrize(
-    "input_config",
-    [
-        None,
-        "input_config_fetch_GIN",
-        "input_config_fetch_local",
-    ],
-)
-def test_entry_point(
-    input_config: Optional[str],
-    monkeypatch: pytest.MonkeyPatch,
-    tmp_path: Path,
-    request: pytest.FixtureRequest,
-):
+def test_entry_point():
     """Test running the cellfinder workflow via the predefined entry point
 
     Parameters
@@ -134,17 +77,9 @@ def test_entry_point(
     request : pytest.FixtureRequest
         Pytest fixture to enable requesting fixtures by name
     """
-    # monkeypatch to change current directory to
-    # pytest temporary directory
-    # (cellfinder cache directory is created in cwd)
-    monkeypatch.chdir(tmp_path)
 
     # define CLI input
-    subprocess_input = ["cellfinder-workflow"]
-    # append config if required
-    if input_config:
-        subprocess_input.append("--config")
-        subprocess_input.append(str(request.getfixturevalue(input_config)))
+    subprocess_input = ["cellfinder-core-workflow"]
 
     # run workflow with no CLI arguments,
     subprocess_output = subprocess.run(
